feat(web.mxplan): remove spf banneer (#10822)
  ref: MANAGER-13036
