+++ packages/manager/modules/email-domain/src/general-informations/email-domain-general-informations.html
@@ -1,5 +1,4 @@
 <div>
-    <wuc-spf-banner></wuc-spf-banner>
     <div data-ovh-alert="{{alerts.main}}"></div>
     <div class="row d-md-flex">
         <div class="col-sm-4 mb-5">
