+++ packages/manager/modules/exchange/src/information/information.html
@@ -1,5 +1,4 @@
 <div>
-    <wuc-spf-banner></wuc-spf-banner>
     <office-attach data-ng-if="ctrl.worldPart === 'EU'"></office-attach>
     <div class="row">
         <div class="col-md-9">
