+++ packages/manager/modules/web-universe-components/src/mxplan-banner/mxplan-banner.html
@@ -1,12 +1,12 @@
 <oui-message class="mb-2" data-type="warning">
-    <span data-translate="spf_banner_info"></span>
+    <span data-translate="mxplan_info"></span>
     <a
         class="oui-link_icon"
         data-ng-href="{{ $ctrl.GUIDE_LINK }}"
         target="_blank"
         rel="noopener"
     >
-        <span data-translate="spf_banner_link"></span>
+        <span data-translate="mxplan_banner_link"></span>
         <i class="oui-icon oui-icon-external-link" aria-hidden="true"></i>
     </a>
 </oui-message>
