+++ packages/manager/modules/web-universe-components/src/mxplan-banner/translations/Messages_de_DE.json
