+++ packages/manager/modules/web-universe-components/src/mxplan-banner/translations/Messages_fr_CA.json
