+++ packages/manager/modules/web-universe-components/src/mxplan-banner/translations/Messages_pl_PL.json
