--- packages/manager/modules/web-universe-components/src/spf-banner/spf-banner.component.js
@@ -1,8 +1,0 @@
-import template from './spf-banner.html';
-import controller from './spf-banner.controller';
-
-export default {
-  name: 'wucSpfBanner',
-  template,
-  controller,
-};
