--- packages/manager/modules/web-universe-components/src/spf-banner/spf-banner.constants.js
@@ -1,32 +1,0 @@
-export const GUIDE_LINKS = {
-  CA:
-    'https://help.ovhcloud.com/csm/en-ca-dns-spf-record?id=kb_article_view&sysparm_article=KB0051715',
-  GB:
-    'https://help.ovhcloud.com/csm/en-gb-dns-spf-record?id=kb_article_view&sysparm_article=KB0051714',
-  DE:
-    'https://help.ovhcloud.com/csm/de-dns-spf-record?id=kb_article_view&sysparm_article=KB0051699',
-  ES:
-    'https://help.ovhcloud.com/csm/es-es-dns-spf-record?id=kb_article_view&sysparm_article=KB0051709',
-  FR:
-    'https://help.ovhcloud.com/csm/fr-dns-spf-record?id=kb_article_view&sysparm_article=KB0051712',
-  IE:
-    'https://help.ovhcloud.com/csm/en-ie-dns-spf-record?id=kb_article_view&sysparm_article=KB0051701',
-  IT:
-    'https://help.ovhcloud.com/csm/it-dns-spf-record?id=kb_article_view&sysparm_article=KB0051721',
-  MA:
-    'https://help.ovhcloud.com/csm/fr-dns-spf-record?id=kb_article_view&sysparm_article=KB0051712',
-  PL:
-    'https://help.ovhcloud.com/csm/pl-dns-spf-record?id=kb_article_view&sysparm_article=KB0051716',
-  PT:
-    'https://help.ovhcloud.com/csm/pt-dns-spf-record?id=kb_article_view&sysparm_article=KB0051717',
-  SN:
-    'https://help.ovhcloud.com/csm/fr-dns-spf-record?id=kb_article_view&sysparm_article=KB0051712',
-  TN:
-    'https://help.ovhcloud.com/csm/fr-dns-spf-record?id=kb_article_view&sysparm_article=KB0051712',
-  DEFAULT:
-    'https://help.ovhcloud.com/csm/en-gb-dns-spf-record?id=kb_article_view&sysparm_article=KB0051714',
-};
-
-export default {
-  GUIDE_LINKS,
-};
