--- packages/manager/modules/web-universe-components/src/spf-banner/translations/Messages_fr_FR.json
@@ -1,4 +1,0 @@
-{
-  "spf_banner_info": "Si vous utilisez l'offre mail d'OVHcloud, nous vous recommandons fortement de vérifier si le SPF de votre zone DNS est bien installé afin de vous assurer de la bonne livraison de vos mails.",
-  "spf_banner_link": "En savoir plus"
-}
